feat: US-05 finish occupied table (#14)

+++ back-end/src/db/migrations/20210204181028_createTablesTable.js
@@ -4,7 +4,7 @@ exports.up = function (knex) {
     table.string("table_name").notNullable();
     table.integer("capacity", null).unsigned().notNullable();
     table
-      .bigInteger("reservation_id")
+      .integer("reservation_id")
       .unsigned()
       .nullable()
       .defaultTo(null)
